New updates

+++ backend/Controllers/CommentController.cs
@@ -64,13 +64,45 @@ namespace backend.Controllers
 
             return Ok(commentModel);
         }
+        [HttpPut]
+        [Route("{id:int}")]
+        [Authorize]
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto updateComment)
+        {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            var username = User.GetUserName();
+            var appUser = await _userManager.FindByNameAsync(username);
+            var comments = await _commentRepo.GetByIdAsync(id);
+            if(appUser.Id != comments.UserId && !User.IsInRole("Admin"))
+            {
+                return BadRequest("You cannot edit other people's comments");
+            }
+            var commentModel = await _commentRepo.UpdateAsync(id,updateComment.ToCommentFromUpdateDto());
+            if(commentModel == null)
+            {
+                return NotFound();
+            }
+            return Ok(commentModel.ToCommentDto());
+
+        }
         [HttpDelete]
         [Route("{id:int}")]
+        [Authorize]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
             if(!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
+            }
+            var username = User.GetUserName();
+            var appUser = await _userManager.FindByNameAsync(username);
+            var comments = await _commentRepo.GetByIdAsync(id);
+            if(appUser.Id != comments.UserId && !User.IsInRole("Admin"))
+            {
+                return BadRequest("You cannot delete other people's comments");
             }
             var comment = await _commentRepo.DeleteAsync(id);
             if(comment == null)
